requirements.txt
asgiref==3.7.2
Django==4.2.3
django-phonenumber-field==7.1.0
phonenumbers==8.13.17
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
pandas == 2.0.2
python-dotenv == 1.0.0
requests==2.31.0